Working wersion. Do not have Readme. :)

+++ src/main/java/karol/train_waybill/front/company/CompanyView.java
@@ -14,6 +14,7 @@ import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.html.Label;
+import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.data.selection.SingleSelect;
 import com.vaadin.flow.router.BeforeEnterEvent;
@@ -24,6 +25,7 @@ import com.vaadin.flow.router.Route;
 import karol.train_waybill.UserDetails.CompanyDetails;
 import karol.train_waybill.database.Company;
 import karol.train_waybill.database.TrainStation;
+import karol.train_waybill.database.TransportStatus;
 import karol.train_waybill.database.Waybill;
 import karol.train_waybill.repository.CompanyRepository;
 import karol.train_waybill.repository.WaybillRepository;
@@ -56,7 +58,7 @@ public class CompanyView extends VerticalLayout implements BeforeEnterObserver {
 		Button add = new Button("Add waybill");
 		
 		add.addClickListener(clickEvent -> {
-			UI.getCurrent().getPage().setLocation("/waybill/add");
+			UI.getCurrent().getPage().setLocation("/company/waybill/add");
 		});
 		
 		add(info, add, edit, delete, grid);
@@ -74,7 +76,8 @@ public class CompanyView extends VerticalLayout implements BeforeEnterObserver {
 		grid.removeColumnByKey("company");
 		
 		grid.addColumn(p -> {
-			return p.getSource_station_id().getFirma();
+			TrainStation ts = p.getSource_station_id();
+			return ts.getFirma();
 		}).setHeader("Stacja nadania");
 		
 		grid.addColumn(p -> {
@@ -91,13 +94,31 @@ public class CompanyView extends VerticalLayout implements BeforeEnterObserver {
 		edit.addClickListener(clickEvent -> {
 			Waybill waybill = waybillSelect.getValue();
 					//grid.getSelectedRow().first();
-			UI.getCurrent().getPage().setLocation("/waybill/edit/" + waybill.getId());
+			if (waybill.getStatus() == TransportStatus.Accept || 
+					waybill.getStatus() == TransportStatus.Report ||
+							waybill.getStatus() == TransportStatus.Reject)
+			{
+				UI.getCurrent().getPage().setLocation("/company/waybill/edit/" + waybill.getId());
+			}
+			else
+			{
+				Notification notification = Notification.show("Edycja listu przewozowego nie jest możliwa!");
+			}
 		});
 		
 		delete.addClickListener(clickEvent -> {
-			Waybill waybill = waybillSelect.getValue();
-					//grid.getSelectedRow().first();
-			waybillRepo.deleteById(waybill.getId());
+			Waybill waybill = waybillSelect.getValue();			
+			if (waybill.getStatus() == TransportStatus.Accept || 
+					waybill.getStatus() == TransportStatus.Report ||
+							waybill.getStatus() == TransportStatus.Reject)
+			{
+				//grid.getSelectedRow().first();
+				waybillRepo.deleteById(waybill.getId());
+			}
+			else
+			{
+				Notification notification = Notification.show("Usunięcie listu przewozowego nie jest możliwe!");
+			}
 		});
 	}
